Change vote weight type and update documentation accordingly

+++ database/migrations/0000_00_00_000000_create_votes_table.php
@@ -14,7 +14,7 @@ class CreateVotesTable extends Migration
             $table->increments('id');
             $table->morphs('voteable');
             $table->morphs('voter');
-            $table->integer('weight');
+            $table->tinyInteger('weight');
             $table->timestamps();
         });
     }
